update loading locator

+++ si-automated-tests/Source/Main/DBModels/PartiesDBModel.cs
@@ -3,7 +3,7 @@ namespace si_automated_tests.Source.Main.DBModels
 {
     public class PartiesDBModel
     {
-        public int creditlimit { get; set; }
+        public Decimal creditlimit { get; set; }
         public int partyID { get; set; }
         public PartiesDBModel()
         {
