+++ si-automated-tests/Source/Main/DBModels/WBPartySettingVDBModel.cs
@@ -3,7 +3,7 @@ namespace si_automated_tests.Source.Main.DBModels
 {
     public class WBPartySettingVDBModel
     {
-        public int creditlimitwarning { get; set; }
+        public decimal creditlimitwarning { get; set; }
         public int partyID { get; set; }
         public WBPartySettingVDBModel()
         {
